% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_utils.R
\name{annotate_dmr}
\alias{annotate_dmr}
\title{Annotate DMRs with overlapping genes}
\usage{
annotate_dmr(dmr_collapsed, gtf_file, feature_type = "gene")
}
\arguments{
\item{dmr_collapsed}{Collapsed DMRs from collapse_dmr}

\item{gtf_file}{Path to GTF annotation file}

\item{feature_type}{Character, feature type to overlap (default "gene")}
}
\value{
data.table with gene annotations added
}
\description{
Annotate DMRs with overlapping genes
}
