% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{apply_kernel_smoothing}
\alias{apply_kernel_smoothing}
\title{Apply kernel smoothing to count matrix}
\usage{
apply_kernel_smoothing(count_matrix, smooth, kernel, kernel_weights)
}
\description{
Apply kernel smoothing to count matrix
}
\keyword{internal}
