% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{apply_smoothing}
\alias{apply_smoothing}
\title{Apply smoothing to count matrix}
\usage{
apply_smoothing(count_matrix, smooth, step)
}
\description{
Apply smoothing to count matrix
}
\keyword{internal}
