% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{calculate_pct_matrix}
\alias{calculate_pct_matrix}
\title{Calculate percentage methylation matrix}
\usage{
calculate_pct_matrix(count_matrix, groups)
}
\description{
Calculate percentage methylation matrix
}
