% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{calculate_percentage_matrix}
\alias{calculate_percentage_matrix}
\title{Calculate percentage matrix from count matrix}
\usage{
calculate_percentage_matrix(count_matrix, groups)
}
\description{
Calculate percentage matrix from count matrix
}
\keyword{internal}
