% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{calculate_read_strategies}
\alias{calculate_read_strategies}
\title{Calculate optimal read strategies based on index}
\usage{
calculate_read_strategies(index)
}
\arguments{
\item{index}{Optimized index object}
}
\value{
List of read strategies by cell
}
\description{
Analyzes data fragmentation to determine the most efficient read strategy
for each cell-chromosome combination.
}
\keyword{internal}
