% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{compare_all_vs_all}
\alias{compare_all_vs_all}
\title{Perform tests of each group against all others}
\usage{
compare_all_vs_all(counts, min_group)
}
\description{
Perform tests of each group against all others
}
\keyword{internal}
