% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_utils.R
\name{configure_dmr_analysis}
\alias{configure_dmr_analysis}
\title{Configure analysis parameters}
\usage{
configure_dmr_analysis(
  genome = "hg38",
  context = "CG",
  window_size = 500,
  smoothing_windows = 3,
  min_coverage = 5
)
}
\arguments{
\item{genome}{Character, genome build}

\item{context}{Character, methylation context}

\item{window_size}{Integer, base window size}

\item{smoothing_windows}{Integer, number of windows to smooth}

\item{min_coverage}{Integer, minimum coverage threshold}
}
\value{
List of analysis parameters
}
\description{
Configure analysis parameters
}
