% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_utils.R
\name{create_dmr_track}
\alias{create_dmr_track}
\title{Create DMR tracks for visualization}
\usage{
create_dmr_track(smoothed_windows, group_col)
}
\arguments{
\item{smoothed_windows}{Smoothed windows from calc_smoothed_windows}

\item{group_col}{Character, which group column to extract}
}
\value{
data.table formatted for genome browser
}
\description{
Create DMR tracks for visualization
}
