% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_utils.R
\name{diagnose_dmr_data}
\alias{diagnose_dmr_data}
\title{Diagnose DMR analysis issues}
\usage{
diagnose_dmr_data(h5_path, barcode, context = "CG")
}
\arguments{
\item{h5_path}{Path to H5 file}

\item{barcode}{Barcode to test}

\item{context}{Methylation context}
}
\value{
List with diagnostic information
}
\description{
Diagnose DMR analysis issues
}
