% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{filter_count_matrix}
\alias{filter_count_matrix}
\title{Filter counts matrix for DMR testing}
\usage{
filter_count_matrix(sum_matrix, min_total)
}
\description{
Filter counts matrix for DMR testing
}
\keyword{internal}
