% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{filter_dmr}
\alias{filter_dmr}
\title{Filter and reshape DMR results}
\usage{
filter_dmr(
  dmr_matrix,
  method = "bonferroni",
  p_threshold = 0.01,
  log_threshold = 1,
  filter = TRUE
)
}
\arguments{
\item{dmr_matrix}{Results from test_dmr}

\item{method}{Multiple testing correction method}

\item{p_threshold}{Maximum adjusted p-value}

\item{log_threshold}{Minimum absolute log fold change}

\item{filter}{Logical, whether to filter results}
}
\value{
data.table with filtered results
}
\description{
Filter and reshape DMR results
}
