% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{generate_genomic_windows}
\alias{generate_genomic_windows}
\title{Generate genomic windows}
\usage{
generate_genomic_windows(genome, step, chrList = NULL, chrSizes = NULL)
}
\description{
Generate genomic windows
}
\keyword{internal}
