% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{load_index}
\alias{load_index}
\title{Load optimized index from disk}
\usage{
load_index(file)
}
\arguments{
\item{file}{Path to the saved index file}
}
\value{
Optimized index object
}
\description{
Load optimized index from disk
}
