% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{merge_with_windows}
\alias{merge_with_windows}
\title{Merge count matrix with genomic windows}
\usage{
merge_with_windows(genomechunks, count_matrix)
}
\description{
Merge count matrix with genomic windows
}
\keyword{internal}
