% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minithyst-package.R
\docType{package}
\name{minithyst-package}
\alias{minithyst}
\alias{minithyst-package}
\title{minithyst: Miniature Amethyst for DMR Analysis from Single-Cell Methylation Data}
\description{
A streamlined implementation of differential methylated region (DMR) analysis from the amethyst package. Provides focused functionality for DMR detection from single-cell methylation data with minimal dependencies. Implements efficient algorithms for indexing H5 files, calculating smoothed methylation windows, and statistical testing for differential methylation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stcolema/minithyst}
  \item Report bugs at \url{https://github.com/stcolema/minithyst/issues}
}

}
\author{
\strong{Maintainer}: Stephen Coleman \email{colemste@ohsu.edu}

}
\keyword{internal}
