% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{print.optimized_methylation_index}
\alias{print.optimized_methylation_index}
\title{Print method for optimized index}
\usage{
\method{print}{optimized_methylation_index}(x, ...)
}
\arguments{
\item{x}{Optimized index object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for optimized index
}
