% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{process_with_optimized_index}
\alias{process_with_optimized_index}
\title{Process data using optimized index}
\usage{
process_with_optimized_index(index, metadata, groups, group_by, step)
}
\description{
Process data using optimized index
}
\keyword{internal}
