% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{reduce_to_real_columns}
\alias{reduce_to_real_columns}
\title{Reduce count matrix to the actual columns removing columns created by merging}
\usage{
reduce_to_real_columns(dt)
}
\description{
Reduce count matrix to the actual columns removing columns created by merging
}
\keyword{internal}
