% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{save_index}
\alias{save_index}
\title{Save optimized index to disk}
\usage{
save_index(index, file)
}
\arguments{
\item{index}{Optimized index object}

\item{file}{Path to save the index (.rds extension recommended)}
}
\description{
Save optimized index to disk
}
