% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_utils.R
\name{summarize_dmr}
\alias{summarize_dmr}
\title{Generate DMR summary statistics}
\usage{
summarize_dmr(dmr_results)
}
\arguments{
\item{dmr_results}{Any DMR result table}
}
\value{
List with summary statistics
}
\description{
Generate DMR summary statistics
}
