% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_windows}
\alias{summarize_windows}
\title{Summarize smoothed window results}
\usage{
summarize_windows(result, type = "sum")
}
\arguments{
\item{result}{Output from calc_smoothed_windows}

\item{type}{Character, either "sum" or "pct" to specify which matrix}
}
\value{
data.table with summary statistics
}
\description{
Provides summary statistics for the output of calc_smoothed_windows.
}
