% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{test_dmr}
\alias{test_dmr}
\title{Test Differentially Methlated Regions}
\usage{
test_dmr(sum_matrix, comparisons = NULL, min_total = 3, min_group = 3)
}
\arguments{
\item{sum_matrix}{data.table with counts from calc_smoothed_windows}

\item{comparisons}{data.frame with test specifications or NULL for all-vs-all}

\item{min_total}{Minimum total observations across groups}

\item{min_group}{Minimum observations per group}
}
\value{
data.table with test results
}
\description{
Test for differentially methylated regions
}
