% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_utils.R
\name{validate_h5_structure}
\alias{validate_h5_structure}
\title{Validate H5 file structure}
\usage{
validate_h5_structure(
  h5_path,
  expected_structure = list(contexts = c("CG", "CH"), required_columns = c("chr", "pos",
    "c", "t"))
)
}
\arguments{
\item{h5_path}{Path to H5 file}

\item{expected_structure}{List describing expected structure}
}
\value{
Logical indicating if structure is valid
}
\description{
Validate H5 file structure
}
