% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr_analysis.R
\name{calc_smoothed_windows}
\alias{calc_smoothed_windows}
\title{Calculate smoothed methylation windows}
\usage{
calc_smoothed_windows(
  h5_paths,
  chr_index,
  metadata,
  type = "CG",
  step = 500,
  smooth = 3,
  group_by = "cluster_id",
  genome = "hg38",
  threads = 1
)
}
\arguments{
\item{h5_paths}{data.table with columns: barcode, path}

\item{chr_index}{List of chromosome indices from index_chromosomes}

\item{metadata}{data.table with cell metadata, must have grouping column}

\item{type}{Character, methylation context}

\item{step}{Integer, window step size in bp}

\item{smooth}{Integer, number of windows to smooth over}

\item{group_by}{Character, column name in metadata for grouping}

\item{genome}{Character, genome build for chromosome sizes}

\item{threads}{Integer, number of threads}
}
\value{
List with sum_matrix and pct_matrix
}
\description{
Calculate smoothed methylation windows
}
